move thie notification load from init to deshboard component
