+++ src/app/services/init.service.ts
@@ -12,9 +12,6 @@ export class InitService {
   constructor(private notificatons:NotificationService, private employees:EmployeeService, private shifts:ShiftService) { }
 
   async doInit(){
-    //this.test();
-    //get all notifications
-    await this.notificatons.getNotifications();
     //get all shifts
     await this.shifts.getshifts();
